+++ NEWS.md
@@ -1,3 +1,9 @@
+# multigroup.vaccine 0.1.2
+
+* Updated `contactMatrixPolymod()` to maintain compatibility with new
+  versions of `socialmixr`, including updated argument names and explicit
+  POLYMOD population data.
+
 # multigroup.vaccine 0.1.1
 
 * Initial release on CRAN